+++ src/Agent/CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### New Features
 * Support of setting up labels via appsettings.json and app/web.config file. [#1204](https://github.com/newrelic/newrelic-dotnet-agent/pull/1204)
 * Additional DEBUG-level logging of all environment variables.
-* Adds a new `SetName()` method to the Agent API for spans that allows customization of segment/span/metric names for custom instrumentation. [#1238](https://github.com/newrelic/newrelic-dotnet-agent/pull/1238)
+* Forwarded application logs now capture exception details including, error message, error stack, and error class. [#1228](https://github.com/newrelic/newrelic-dotnet-agent/pull/1228)
+  * Log events with no message will now be accepted if an exception is present in the log event.
+  * The error stack is created using the stack of the inner exception, up to 5 levels deep, just like existing Agent error reporting.
+* Adds a new `SetName()` method to the Agent API for spans which allows customization of segment/span/metric names. [#1238](https://github.com/newrelic/newrelic-dotnet-agent/pull/1238)
 
 ### Fixes
 * Resolves an issue where log forwarding could drop logs in async scenarios. [#1201](https://github.com/newrelic/newrelic-dotnet-agent/pull/1201)